+++ app/src/main/java/io/taptap/stupidenglish/ui/ChooseGroupBottomSheet.kt
@@ -11,9 +11,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import io.taptap.stupidenglish.R
-import io.taptap.stupidenglish.base.logic.sources.groups.read.GroupItemUI
-import io.taptap.stupidenglish.base.logic.sources.groups.read.GroupListModels
-import io.taptap.stupidenglish.base.logic.sources.groups.read.NoGroup
+import io.taptap.uikit.group.GroupItemUI
+import io.taptap.uikit.group.GroupListModels
+import io.taptap.uikit.group.NoGroup
 import io.taptap.uikit.AverageTitle
 import io.taptap.uikit.BottomSheetScreen
 import io.taptap.uikit.PrimaryButton
