+++ app/src/main/res/values-ru/strings.xml
@@ -44,7 +44,6 @@
     <string name="word_get_groups_error">Возникла ошибка в момент получения групп</string>
     <string name="word_group_title">Группы</string>
     <string name="word_group_button">Добавить</string>
-    <string name="word_group_no_group_name">Все слова</string>
     <string name="word_group_add_group_title">Новая группа</string>
     <string name="word_empty_list_title">Список пуст</string>
     <string name="word_empty_list_description">Добавь слова в эту группу</string>
