+++ app/src/main/res/values/strings.xml
@@ -33,7 +33,6 @@
     <string name="word_fab_text">Add word</string>
     <string name="word_group_title">Groups</string>
     <string name="word_group_button">Add</string>
-    <string name="word_group_no_group_name">All words</string>
     <string name="word_group_add_group_title">New group</string>
     <string name="word_empty_list_title">List is empty</string>
     <string name="word_empty_list_description">Add some words to the group</string>
