+++ core/uikit/src/main/java/io/taptap/uikit/group/GroupItemHeader.kt
@@ -1,4 +1,4 @@
-package io.taptap.stupidenglish.ui
+package io.taptap.uikit.group
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
