+++ core/uikit/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="word_group_no_group_name">All words</string>
+</resources>
